Fix circular reference
