+++ cstar/orchestration/adapter.py
@@ -1,5 +1,6 @@
 import typing as t
 
+from cstar.base.adapter import ModelAdapter
 from cstar.base.additional_code import AdditionalCode
 from cstar.execution.file_system import RomsFileSystemManager
 from cstar.marbl.external_codebase import MARBLExternalCodeBase
@@ -18,28 +19,6 @@ from cstar.roms.input_dataset import (
     ROMSSurfaceForcing,
     ROMSTidalForcing,
 )
-
-_Tin = t.TypeVar("_Tin")
-_Tout_co = t.TypeVar("_Tout_co", covariant=True)
-
-
-class ModelAdapter(t.Protocol, t.Generic[_Tin, _Tout_co]):
-    """Contract exposing a mechanism to adapt a source model to a target type."""
-
-    model: _Tin
-
-    def __init__(self, model: _Tin) -> None:
-        self.model = model
-
-    def adapt(self) -> _Tout_co | None:
-        """Adapt the source model to the target output type.
-
-        Returns
-        -------
-        _Tout
-            The instance converted from the source model
-        """
-        ...
 
 
 class DiscretizationAdapter(
